BSM 0.2; Condition Sets, Hotkeys & Editable Vars
Added ability to have 'ConditionSets'.  Condition sets allow user to
change the behavior of the state machine on the fly, essentially putting
it into a different 'state'.  Effectively, this is done by setting the
'CurrentConditionSet' variable.  This variable can be used flexibly (to
flag different transitions, change condition selection, etc).  Toggle
buttons to set the condition set are only displayed if NumConditionSets
is greater than 1.

Added ability to have 'hotkeys' available to the user.  Currently
supports up to 5 hotkeys.  Hotkey presses are executed at the end of the
trial (in the ITI).  As they just execute a function call they can be
used to change variables, plot figures, etc.  Buttons for hotkeys are
only shown if at least one is defined.

Added 'Editable' optional flag to condition variables to determine
whether they are editable online.  Default is true, but if set to false
then the variable will not be available to user to change during
execution.

Added estimated end time to the GUI.

Updated BSM version number as the Read/Write functions are no longer
backwards compatible.